Cypress:/E2E Test Fixed
