+++ e2e/cypress/tests/integration/accessibility/accessibility_post_spec.js
@@ -183,7 +183,7 @@ describe('Verify Accessibility Support in Post', () => {
                 // eslint-disable-next-line no-negated-condition
                 if (!emojiPickerEnabled) {
                     // * Verify focus is on the actions button
-                    cy.get(`#CENTER_button_${postId}`).should('be.focused').and('have.attr', 'aria-label', 'more actions');
+                    cy.get(`#CENTER_button_${postId}`).should('be.focused').and('have.attr', 'aria-label', 'more');
                     cy.focused().tab();
                 } else {
                     for (let i = 0; i < 3; i++) {
@@ -207,7 +207,7 @@ describe('Verify Accessibility Support in Post', () => {
 
                 if (emojiPickerEnabled) {
                     // * Verify focus is on the actions button
-                    cy.get(`#CENTER_button_${postId}`).should('be.focused').and('have.attr', 'aria-label', 'more actions');
+                    cy.get(`#CENTER_button_${postId}`).should('be.focused').and('have.attr', 'aria-label', 'more');
                     cy.focused().tab();
                 }
 
@@ -247,7 +247,7 @@ describe('Verify Accessibility Support in Post', () => {
 
                 if (emojiPickerEnabled) {
                     // * Verify focus is on the actions button
-                    cy.get(`#RHS_COMMENT_button_${postId}`).should('be.focused').and('have.attr', 'aria-label', 'more actions');
+                    cy.get(`#RHS_COMMENT_button_${postId}`).should('be.focused').and('have.attr', 'aria-label', 'more');
                     cy.focused().tab({shift: true});
                 }
 
@@ -262,7 +262,7 @@ describe('Verify Accessibility Support in Post', () => {
                 // eslint-disable-next-line no-negated-condition
                 if (!emojiPickerEnabled) {
                     // * Verify focus is on the actions button
-                    cy.get(`#RHS_COMMENT_button_${postId}`).should('be.focused').and('have.attr', 'aria-label', 'more actions');
+                    cy.get(`#RHS_COMMENT_button_${postId}`).should('be.focused').and('have.attr', 'aria-label', 'more');
                     cy.focused().tab({shift: true});
                 } else {
                     cy.get('#recent_reaction_0').should('have.class', 'emoticon--post-menu');
